+++ CHANGES.md
@@ -1,16 +1,49 @@
 # Changes
 
+### Add
+
+* `QueryConfig`, similar to `JsonConfig` for customizing error handling of query extractors.
+
+### Changes
+
+* `JsonConfig` is now `Send + Sync`, this implies that `error_handler` must be `Send + Sync` too.
+
+## [1.0.0-beta.4] - 2019-05-12
+
+### Add
+
+* Allow to set/override app data on scope level
+
+### Changes
+
+* `App::configure` take an `FnOnce` instead of `Fn`
+* Upgrade actix-net crates
+
+## [1.0.0-beta.3] - 2019-05-04
+
 ### Added
 
 * Add helper function for executing futures `test::block_fn()`
 
 ### Changed
 
+* Extractor configuration could be registered with `App::data()`
+  or with `Resource::data()` #775
+
+* Route data is unified with app data, `Route::data()` moved to resource
+  level to `Resource::data()`
+
+* CORS handling without headers #702
+
 * Allow to construct `Data` instances to avoid double `Arc` for `Send + Sync` types.
 
 ### Fixed
 
 * Fix `NormalizePath` middleware impl #806
+
+### Deleted
+
+* `App::data_factory()` is deleted.
 
 
 ## [1.0.0-beta.2] - 2019-04-24